bottom up partitioning

+++ src/vec.c
@@ -1,6 +1,6 @@
 #include "vec.h"
 
-int _vec_expand(char **data, unsigned *length, unsigned *capacity, unsigned memsz){
+int _vec_expand(char **data, unsigned long *length, unsigned long *capacity, unsigned long memsz){
 	if(*length + 1 > *capacity){
 		void *ptr;
 		int n = (*capacity == 0) ? 1 : *capacity << 1;
